requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.2
